backend/requirements.txt
Flask==2.0.1
flask-cors==3.0.10
PyPDF2==3.0.1
python-docx==0.8.11
SQLAlchemy==1.4.23
python-dotenv==0.19.0
werkzeug==2.0.1